Add gaussian stochastic correction
The training data generator now includes stochastic variance along a bell curve.  With a curve that allows a 75%-125% multiplier to a creature's statistics (to simulate environmental influence on the expression of genes), the system was able to achieve+/- 6.21% accuracy, indicating that even when the groundtruth data is itself noisy, the network can still detect the causal relationship between genotype and phenotype.